keep all of the rendering of kernel methods in acc.kernels. some handling was in homogeneous_single_scatterer
